Refactored to have generic timeout wrapper
