+++ CMakeLists.txt
@@ -33,7 +33,10 @@ add_library(
     ${CMAKE_CURRENT_SOURCE_DIR}/include/foxy/session.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/include/foxy/shared_handler_ptr.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/include/foxy/utility.hpp
+
     ${CMAKE_CURRENT_SOURCE_DIR}/include/foxy/detail/export_non_connect_fields.hpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/include/foxy/detail/timed_op_wrapper.hpp
+
     ${CMAKE_CURRENT_SOURCE_DIR}/include/foxy/impl/session.impl.hpp
     ${CMAKE_CURRENT_SOURCE_DIR}/include/foxy/impl/shared_handler_ptr.impl.hpp
